refactor to separate main stack from lsdome geometry
