+++ src/main/java/me/lsdo/processing/DomePixel.java
@@ -2,7 +2,7 @@ package me.lsdo.processing;
 
 // Representation of a coordinate within the dome pixel grid. See TriCoord.java for more info.
 
-public class DomeCoord {
+public class DomePixel extends LedPixel {
 
     // Every pixel in the dome has a unique coordinate.
     public TriCoord universal;
@@ -13,13 +13,13 @@ public class DomeCoord {
     // and a within-panel pixel component.
     public TriCoord pixel;
 
-    public DomeCoord(TriCoord uni, int panel_length) {
+    public DomePixel(TriCoord uni, int panel_length) {
         universal = uni;
         panel = TriCoord.toPanel(uni, panel_length);
         pixel = TriCoord.toPixel(uni, panel_length);
     }
 
-    public DomeCoord(TriCoord panel, TriCoord pixel) {
+    public DomePixel(TriCoord panel, TriCoord pixel) {
         universal = TriCoord.toUniversal(panel, pixel);
         this.panel = panel;
         this.pixel = pixel;
@@ -34,9 +34,13 @@ public class DomeCoord {
         }
     }
 
+    public PVector2 toXY() {
+	throw new RuntimeException();
+    }
+    
     public boolean equals(Object o) {
-        if (o instanceof DomeCoord) {
-            DomeCoord dc = (DomeCoord)o;
+        if (o instanceof DomePixel) {
+            DomePixel dc = (DomePixel)o;
             return universal.equals(dc.universal);
         } else {
             return false;
