+++ src/main/java/me/lsdo/processing/PixelGridSketch.java
@@ -16,17 +16,17 @@ import processing.core.PApplet;
 public class PixelGridSketch {
 
     protected PApplet app;
-    protected DomeAnimation animation;
+    protected DomeAnimation<? extends LedPixel> animation;
 
-    private Map<DomeCoord, PVector2> screenCoords;
+    private Map<LedPixel, PVector2> screenCoords;
     
-    public PixelGridSketch(PApplet app, DomeAnimation animation) {
+    public PixelGridSketch(PApplet app, DomeAnimation<? extends LedPixel> animation) {
         this.app = app;
         this.animation = animation;
 
 	// Pre-compute screen coordinates for dome pixels.
-	screenCoords = new HashMap<DomeCoord, PVector2>();
-        for (DomeCoord c : animation.dome.coords) {
+	screenCoords = new HashMap<LedPixel, PVector2>();
+        for (LedPixel c : animation.dome.coords) {
             PVector2 screenCoord = animation.dome.domeCoordToScreen(c, app.width, app.height);
 	    screenCoords.put(c, screenCoord);
         }
@@ -39,14 +39,14 @@ public class PixelGridSketch {
 
         app.background(0);
         app.noStroke();
-        for (DomeCoord c : animation.dome.coords){
+        for (LedPixel c : animation.dome.coords){
             PVector2 p = screenCoords.get(c);
             app.fill(animation.dome.getColor(c));
             app.ellipse(p.x, p.y, 3, 3);
         }
 
         app.fill(127);
-        app.text("opc @" + animation.getOpcHost(), 100, app.height - 10);
+        app.text("opc @" + animation.dome.getOpcHosts(), 100, app.height - 10);
         app.text(String.format("%.1ffps", app.frameRate), 10, app.height - 10);
 
     }
